Mark `GraphMakie` and `Graphs` as weak-dependency
Closes #125
